% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Get Available \pkg{lexicon} Data}
\usage{
available_data()
}
\value{
Returns a data.frame
}
\description{
See available \pkg{lexicon} data a data.frame.
}
\examples{
available_data()
}
