% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_word_list.R
\docType{data}
\name{enable_word_list}
\alias{enable_word_list}
\title{ENABLE Word List}
\format{A character vector with 172,823 elements}
\usage{
data(enable_word_list)
}
\description{
A dataset containing a character vector of words from the Enhanced North
American Benchmark Lexicon (ENABLE) which is used in the game Words With
Friends (\url{https://en.wikipedia.org/wiki/Words_with_Friends}).  This
version of the list comes from \url{http://www.puzzlers.org/pub/wordlists}.
}
\references{
\url{http://www.puzzlers.org/pub/wordlists}
}
\keyword{datasets}
