% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_first_names.R
\docType{data}
\name{freq_first_names}
\alias{freq_first_names}
\title{Frequent U.S. First Names}
\format{A data frame with 5494 rows and 4 variables}
\usage{
data(freq_first_names)
}
\description{
A dataset containing frequent first names based on the 1990 U.S. census.
}
\details{
\itemize{
  \item Name. A first name
  \item n. The approximate frequency within the sex
  \item prop. The proportion within the sex
  \item sex. The sex corresponding to the name
}
}
\references{
http://names.mongabay.com
}
\keyword{datasets}
