% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_last_names.R
\docType{data}
\name{freq_last_names}
\alias{freq_last_names}
\title{Frequent U.S. Last Names}
\format{A data frame with 14,840 rows and 3 variables}
\usage{
data(freq_last_names)
}
\description{
A dataset containing frequent last names based on the 1990 U.S. census.
}
\details{
\itemize{
  \item Surname. A last name
  \item n. The approximate frequency
  \item prop. The proportion
}
}
\references{
http://names.mongabay.com
}
\keyword{datasets}

