% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_contractions.R
\docType{data}
\name{hash_contractions}
\alias{hash_contractions}
\title{Contraction Conversions}
\format{A data frame with 70 rows and 2 variables}
\usage{
data(hash_contractions)
}
\description{
A dataset containing common contractions and their expanded form.
}
\details{
\itemize{
  \item contraction. The contraction word
  \item expanded. The expanded form of the contraction
}
}
\keyword{datasets}

