% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_emojis.R
\docType{data}
\name{hash_emojis_identifier}
\alias{hash_emojis_identifier}
\title{Emoji Identifier Lookup Table}
\format{A data frame with 734 rows and 2 variables}
\usage{
data(hash_emojis_identifier)
}
\description{
A dataset containing ASCII byte code representation of emojis and their
accompanying identifier (for use in the \pkg{textclean} or \pkg{sentimentr}
packages).
}
\details{
\itemize{
  \item x. Byte code representation of emojis
  \item y. Emoji description
}
}
\references{
\url{http://www.unicode.org/emoji/charts/full-emoji-list.html}
}
\keyword{datasets}
