% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_emoticons.R
\docType{data}
\name{hash_emoticons}
\alias{hash_emoticons}
\title{Emoticons}
\format{A data frame with 75 rows and 2 variables}
\usage{
data(hash_emoticons)
}
\description{
A \pkg{data.table} key containing common emoticons (adapted from
\href{http://www.lingo2word.com/lists/emoticon_listH.html}{Popular Emoticon List}).
}
\details{
\itemize{
  \item x. The graphic representation of the emoticon
  \item y. The meaning of the emoticon
}
}
\examples{
\dontrun{
library(data.table)
hash_emoticons[c(':-(', '0;)')]
}
}
\references{
\url{http://www.lingo2word.com/lists/emoticon_listH.html}
}
\keyword{datasets}
