% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_grade.R
\docType{data}
\name{hash_grade}
\alias{hash_grade}
\title{Grades Hash}
\format{A data frame with 15 rows and 2 variables}
\usage{
data(hash_grade)
}
\description{
A dataset containing letter grades and corresponding semantic meaning.
}
\details{
\itemize{
  \item x. Letter grade
  \item y. Semantic meaning of grade
}
}
\keyword{datasets}

