% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_nrc_emotion.R
\docType{data}
\name{hash_nrc_emotions}
\alias{hash_nrc_emotions}
\title{NRC Emotion Table}
\format{A data frame with 8265 rows and 2 variables}
\usage{
data(hash_nrc_emotions)
}
\description{
A \pkg{data.table} dataset containing a filtered version of Mohammad &
Turney', P. D.'s (2010) emotion word list as lookup table.
}
\details{
\itemize{ 
  \item token. A search token indicating emotion.
  \item emotion. An accompanying emotion assocatiated with the token.
}
}
\keyword{datasets}
