% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_power.R
\docType{data}
\name{hash_power}
\alias{hash_power}
\title{Power Lookup Key}
\format{A data frame with 872 rows and 2 variables}
\usage{
data(hash_power)
}
\description{
A \pkg{data.table} containing a power lookup key.
}
\details{
\itemize{
  \item x. A power word
  \item y. A positive or negative value indicating the direction of power in relation to the subject
}
}
\examples{
\dontrun{
library(data.table)
hash_power[c('yield', 'admonish', 'abdicate')]
}
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}
