% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_inquirer.R
\docType{data}
\name{hash_sentiment_inquirer}
\alias{hash_sentiment_inquirer}
\title{Inquirer Polarity Lookup Table}
\format{A data frame with 3,450 rows and 2 variables}
\usage{
data(hash_sentiment_inquirer)
}
\description{
A \pkg{data.table} dataset containing an augmented version of General Inquirer's
positive/negative word list as sentiment lookup values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment
}
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/homecat.htm}
}
\keyword{datasets}
