% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_jockers.R
\name{hash_sentiment_jockers}
\alias{hash_sentiment_jockers}
\title{Jockers Polarity Lookup Table}
\usage{
hash_sentiment_jockers
}
\description{
A \pkg{data.table} dataset containing a modified version of Jocker's (2017)
sentiment lookup table used in \pkg{syuzhet}.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values ranging between -1 and 1.
}
}
\references{
Jockers, M. L. (2017). Syuzhet: Extract sentiment and plot arcs
from Text. Retrieved from https://github.com/mjockers/syuzhet
}
\keyword{datasets}
