% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_slangsd.R
\docType{data}
\name{hash_sentiment_slangsd}
\alias{hash_sentiment_slangsd}
\title{SlangSD Sentiment Polarity Table}
\format{
A data frame with 48,277 rows and 2 variables
}
\usage{
data(hash_sentiment_slangsd)
}
\description{
A \pkg{data.table} dataset containing a filtered version of Wu, Morstatter, 
& Liu's (2016) positive/negative slang word list as sentiment lookup values.
All words containing other than \code{"[a-z ']"} have been removed as well as
any neutral words.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values (+1, -1)
}

Original Licensing:
The dictionary is free to use. If you use it for an academic publication, we 
ask that you cite it using the citation below. If it is used in anything other 
than an academic publication, we ask that you provide a credit and link to SlangSD.com.

article\{DBLP:journals/corr/Wu-etal16,
  author    = \{Liang Wu and Fred Morstatter and Huan Liu\},
  title     = \{SlangSD: Building and Using a Sentiment Dictionary of Slang 
               Words for Short-Text Sentiment Classification\},
  journal   = \{CoRR\},
  volume    = \{abs/1608.05129\},
  year      = \{2016\},
  url       = \{http://arxiv.org/abs/1608.05129\},
  timestamp = \{Wed, 17 Aug 2016 23:32:57 GMT\}
\}
}
\references{
Wu, L., Morstatter, F.,  and Liu, H. (2016). SlangSD: Building 
and using a sentiment dictionary of slang words for short-text sentiment 
classification.  CoRR. abs/1168.1058.  1-15.  \cr \cr
http://slangsd.com
}
\keyword{datasets}
