% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_strength.R
\docType{data}
\name{hash_strength}
\alias{hash_strength}
\title{Strength Lookup Key}
\format{A data frame with 2085 rows and 2 variables}
\usage{
data(hash_strength)
}
\description{
A \pkg{data.table} containing a strength lookup key.
}
\details{
\itemize{
  \item x. A power word
  \item y. A positive or negative value indicating the direction of strength in relation to the subject
}
}
\examples{
\dontrun{
library(data.table)
hash_strength[c('yield', 'admonish', 'abdicate')]
}
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

