% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_abbreviation.R
\docType{data}
\name{key_abbreviation}
\alias{key_abbreviation}
\title{Common Abbreviations}
\format{A data frame with 138 rows and 2 variables}
\usage{
data(key_abbreviation)
}
\description{
A dataset containing a hash lookup of common abbreviations and their long
form.
}
\details{
\itemize{
  \item abbreviation. An abbreviation
  \item phrase. The equivalent word/phrase
}
}
\references{
\url{http://public.oed.com/how-to-use-the-oed/abbreviations}
}
\keyword{datasets}

