% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_power.R
\docType{data}
\name{key_power}
\alias{key_power}
\title{Power Lookup Key}
\format{A data frame with 872 rows and 2 variables}
\usage{
data(key_power)
}
\description{
A \pkg{data.table} containing a power lookup key.
}
\details{
\itemize{
  \item x. A power word
  \item y. A positive or negative value indicating the direction of power in relation to the subject
}
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

