% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_sentiment.R
\docType{data}
\name{key_sentiment}
\alias{key_sentiment}
\title{Sentiment Lookup Key}
\format{A data frame with 6827 rows and 2 variables}
\usage{
data(key_sentiment)
}
\description{
A \pkg{data.table} dataset containing an augmented version of Hu & Liu's (2004)
positive/negative word list as sentiment lookup values.
}
\details{
\itemize{
  \item x. Words
  \item y. Sentiment values (+1, -1)
}
}
\examples{
\dontrun{
library(data.table)
key_sentiment[c('happy', 'angry')]
}
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in customer
reviews. National Conference on Artificial Intelligence.
}
\keyword{datasets}

