% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_sentiment_jockers.R
\name{key_sentiment_jockers}
\alias{key_sentiment_jockers}
\title{Jockers Sentiment Key}
\usage{
key_sentiment_jockers
}
\description{
A dataset containing an imported version of Jocker's (2017)
sentiment lookup table used in \pkg{syuzhet}.
}
\details{
\itemize{
  \item word. Words
  \item value. Sentiment values ranging between -1 and 1.
}
}
\references{
Jockers, M. L. (2017). Syuzhet: Extract sentiment and plot arcs
from Text. Retrieved from https://github.com/mjockers/syuzhet
}
\keyword{datasets}
