% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_strength.R
\docType{data}
\name{key_strength}
\alias{key_strength}
\title{Strength Lookup Key}
\format{A data frame with 2085 rows and 2 variables}
\usage{
data(key_strength)
}
\description{
A \pkg{data.table} containing a strength lookup key.
}
\details{
\itemize{
  \item x. A power word
  \item y. A positive or negative value indicating the direction of strength in relation to the subject
}
}
\references{
\url{http://www.wjh.harvard.edu/~inquirer/inqdict.txt}
}
\keyword{datasets}

