% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_syllable.R
\docType{data}
\name{key_syllable}
\alias{key_syllable}
\title{Syllable Counts}
\format{A data frame with 124603 rows and 2 variables}
\usage{
data(key_syllable)
}
\description{
A \pkg{data.table} hash table dataset containing words and syllable counts.
}
\details{
\itemize{
  \item word. A character coolumn of lower case words.
  \item syllables. The syllable counts per word.
}
}
\examples{
\dontrun{
library(data.table)
key_syllable[c('yield', 'hurtful', 'admonishing', 'abdicate')]
}
}
\references{
Counts scraped from \url{http://www.poetrysoup.com}
}
\keyword{datasets}

