% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrc_emotions.R
\docType{data}
\name{nrc_emotions}
\alias{nrc_emotions}
\title{NRC Emotions}
\format{A data frame with 14182 rows and 9 variables}
\usage{
data(nrc_emotions)
}
\description{
A \pkg{data.table} dataset containing Mohammad & Turney', P. D.'s (2010)
emotions word list as a binary table.
}
\details{
\itemize{
  \item term. A term
  \item anger. Counts of anger anger
  \item anticipation. Counts of anticipation
  \item disgust. Counts of dsgust
  \item fear. Counts of fear
  \item joy. Counts of joy
  \item sadness. Counts of sadness
  \item surprise. Counts of surprise
  \item trust. Counts of trust
}
}
\references{
http://www.purl.com/net/lexicons  \cr \cr
Mohammad, S. M. & Turney, P. D. (2010) Emotions evoked by common words and
phrases: Using Mechanical Turk to create an emotion lexicon, In Proceeding of
Workshop on Computational Approaches to Analysis and Generation of Emotion in
Text, 26-34.
}
\keyword{datasets}

