% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_df_irregular_nouns.R
\docType{data}
\name{pos_df_irregular_nouns}
\alias{pos_df_irregular_nouns}
\title{Irregular Nouns Word Dataframe}
\format{A data frame with 106 rows and 2 variables}
\usage{
data(pos_df_irregular_nouns)
}
\description{
A dataset containing a \code{data.frame} of irregular noun singular and 
plural forms.
}
\details{
\itemize{ 
  \item singular. The singular form of the noun
  \item plural. The plural form of the noun
}
}
\references{
\url{http://www.esldesk.com/vocabulary/irregular-nouns}
}
\keyword{datasets}
