% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_unchanging_nouns.R
\docType{data}
\name{pos_unchanging_nouns}
\alias{pos_unchanging_nouns}
\title{Nouns that are the Same Plural/Singular}
\format{A character vector with 95 elements}
\usage{
data(pos_unchanging_nouns)
}
\description{
A dataset containing a character vector of nouns that have a single form
for both singular and plural (or a singular/plural form does not exist).
}
\details{
These are a subset of irreguar nouns that are: plurale tantum,
singularia tantum, or unchanging.
}
\references{
\url{https://www.vappingo.com/word-blog/101-words-that-are-both-plural-and-singular}
}
\keyword{datasets}
