% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_banned.R
\docType{data}
\name{profanity_banned}
\alias{profanity_banned}
\title{bannedwordlist.com's List of Profane Words}
\format{A character vector with 77 elements}
\usage{
data(profanity_banned)
}
\description{
A dataset containing a character vector of profane words from bannedwordlist.com.
}
\references{
\url{http://www.bannedwordlist.com}
}
\keyword{datasets}
