% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_google.R
\docType{data}
\name{profanity_google}
\alias{profanity_google}
\title{Google's List of Profane Words}
\format{A character vector with 451 elements}
\usage{
data(profanity_google)
}
\description{
A dataset containing a character vector of profane words from Google's
"what do you love" project, compiled by Jamie Wilkinson.
}
\references{
\url{https://gist.github.com/jamiew/1112488}
}
\keyword{datasets}
