% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profanity_von_ahn.R
\docType{data}
\name{profanity_von_ahn}
\alias{profanity_von_ahn}
\title{Luis von Ahn's List of Profane Words}
\format{A character vector with 1384 elements}
\usage{
data(profanity_von_ahn)
}
\description{
A dataset containing a character vector of profane words from Luis von Ahn's
research group.
}
\references{
\url{http://www.cs.cmu.edu/~biglou/resources}
}
\keyword{datasets}
