% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_lucene.R
\docType{data}
\name{sw_lucene}
\alias{sw_lucene}
\title{Lucene Stopword List}
\format{
A character vector with 33 elements
}
\usage{
data(sw_lucene)
}
\description{
A dataset containing a character vector of Lucene's stopwords used in
\code{StopAnalyzer.ENGLISH_STOP_WORDS_SE}.
}
\details{
Lucene's License:

Licensed to the Apache Software Foundation (ASF) under one or more
contributor license agreements.  See the NOTICE file distributed with
this work for additional information regarding copyright ownership.
The ASF licenses this file to You under the Apache License, Version 2.0
(the "License"); you may not use this file except in compliance with
the License.  You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
}
\references{
\url{https://lucene.apache.org/core/4_0_0/analyzers-common/org/apache/lucene/analysis/core/StopFilter.html}
}
\keyword{datasets}
