% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_mallet.R
\docType{data}
\name{sw_mallet}
\alias{sw_mallet}
\title{MALLET Stopword List}
\format{A character vector with 523 elements}
\usage{
data(sw_mallet)
}
\description{
A stopword list containing a character vector of stopwords.
}
\details{
\href{http://mallet.cs.umass.edu/}{From MAchine Learning for LanguagE Toolkit}
}
\references{
\url{http://mallet.cs.umass.edu}
}
\keyword{datasets}
