% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_lemmas.R
\docType{data}
\name{hash_lemmas}
\alias{hash_lemmas}
\title{Lemmatization List}
\format{A data frame with 41,533 rows and 2 variables}
\usage{
data(hash_lemmas)
}
\description{
A dataset based on M\u{e}chura's (2016) English lemmatization list.  This
data set can be useful for join style lemma replacement of inflected token
forms to their root lemmas.  While this is not a true morphalogical analysis
this style of lemma replacement is fast and typically still robust.
}
\details{
\itemize{
  \item token. An inflected token with affixes
  \item lemma. A base form
}
}
\references{
M\u{e}chura, M. B. (2016). \emph{Lemmatization list: English (en)} [Data file]. Retrieved from \url{http://www.lexiconista.com}
}
\keyword{datasets}

