% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon-package.R
\docType{package}
\name{lexicon}
\alias{lexicon}
\alias{lexicon-package}
\alias{package-lexicon}
\title{Lexicons for Text Analysis}
\description{
A collection of lexical hash tables, dictionaries, and word lists.
}

