% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_python.R
\docType{data}
\name{sw_python}
\alias{sw_python}
\title{Python Stopword List}
\format{A character vector with 174 elements}
\usage{
data(sw_python)
}
\description{
A dataset containing a character vector of Python's stopwords.
}
\details{
Copyright (c) 2014, Alireza Savand, Contributors
All rights reserved.
}
\references{
\url{https://pypi.python.org/pypi/stop-words}
}
\keyword{datasets}
